Add files via upload
This project has a lot of components to implement. It is a social media project. We all know about Instagram, Twitter, Facebook, etc.   Now I can create like this. This social media is called "BlinkMedia".  It has a lot of features like relationship (finding the person), and contact with other people.  Users can post it or like and share also followed.